fix IPCam

+++ IPCam.py
@@ -33,7 +33,7 @@ class IPCam:
             self.img = cv2.resize(img, [1280, 720], interpolation = cv2.INTER_AREA)
             self.last_frame = time.time()
 
-            cv2.waitKey(1/self.fps)
+            cv2.waitKey(int(1/self.fps))
 
     def getentrance_name(self):
         return self.entrance_name
